LJPS-53 updated unit test for delete learning journey

+++ backend/service/src/LearningJourneyService.js
@@ -21,6 +21,9 @@ function learningJourneyService(database){
   }
 
   learningJourneyService.deleteLearningJourney = async (learningJourneyId) => {
+    if(!learningJourneyId){
+      return false 
+    }
     try{
       return await database.deleteLearningJourney(learningJourneyId)
     }
